+++ src/TestTrackingDiagrams.Extensions.PubSub/PubSubOperation.cs
@@ -1,0 +1,14 @@
+namespace TestTrackingDiagrams.Extensions.PubSub;
+
+public enum PubSubOperation
+{
+    Publish,
+    PublishBatch,
+    Pull,
+    Acknowledge,
+    ModifyAckDeadline,
+    Receive,
+    StartSubscriber,
+    StopSubscriber,
+    Other
+}