+++ src/TestTrackingDiagrams.Extensions.PubSub/PubSubTrackingVerbosity.cs
@@ -1,0 +1,3 @@
+namespace TestTrackingDiagrams.Extensions.PubSub;
+
+public enum PubSubTrackingVerbosity { Raw, Detailed, Summarised }